refactor: IntegrationFixture TestFixture 로 네임 변경
