+++ src/test/java/subway/integration/LineIntegrationTest.java
@@ -4,14 +4,14 @@ import static io.restassured.RestAssured.given;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.tuple;
 import static org.junit.jupiter.api.Assertions.assertAll;
-import static subway.integration.IntegrationFixture.GANGNAM;
-import static subway.integration.IntegrationFixture.JAMSIL;
-import static subway.integration.IntegrationFixture.LINE_2;
-import static subway.integration.IntegrationFixture.LINE_3;
-import static subway.integration.IntegrationFixture.OBJECT_MAPPER;
-import static subway.integration.IntegrationFixture.SAMSUNG;
-import static subway.integration.IntegrationFixture.SEONGLENUG;
-import static subway.integration.IntegrationFixture.jsonSerialize;
+import static subway.integration.TestFixture.GANGNAM;
+import static subway.integration.TestFixture.JAMSIL;
+import static subway.integration.TestFixture.LINE_2;
+import static subway.integration.TestFixture.LINE_3;
+import static subway.integration.TestFixture.OBJECT_MAPPER;
+import static subway.integration.TestFixture.SAMSUNG;
+import static subway.integration.TestFixture.SEONGLENUG;
+import static subway.integration.TestFixture.jsonSerialize;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import io.restassured.RestAssured;
